requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.4
cffi==1.11.5
cryptography==2.3
Fabric3==1.14.post1
gunicorn==19.9.0
idna==2.7
invoke==1.1.0
paramiko==2.4.1
psycopg2==2.7.5
pyasn1==0.4.4
pycparser==2.18
PyNaCl==1.2.1
pytz==2018.5
six==1.11.0
